requirements.txt
streamlit
diffusers
torch
transformers
Pillow
accelerate